+++ django_boost/models/fields/__init__.py
@@ -1,27 +1,11 @@
-from warnings import warn
-
 from django import forms
 from django.db import models
-from django.db.models.fields import CharField, TextField
+from django.db.models.fields import CharField
 
 from django_boost.forms import fields
 from django_boost.models.fields.related_descriptors import (
     AutoReverseOneToOneDescriptor)
-from django_boost.validators import validate_color_code, validate_json
-
-
-class JsonField(TextField):
-    """TextField with json validation."""
-
-    empty_strings_allowed = False
-    default_validators = [validate_json]
-
-    def __init__(self, *args, **kwargs):
-        warn("django_boost JsonField is deprecated and will be removed in "
-             "django-boost 3.0. Use django.db.models.JSONField instead "
-             "(available since Django 3.1).",
-             DeprecationWarning, stacklevel=2)
-        super().__init__(*args, **kwargs)
+from django_boost.validators import validate_color_code
 
 
 class ColorCodeFiled(CharField):